Fix and add more
